fix for broken memrise pool info (fixes #84)
Strange things may happen in Memrise world... in this case the pool id of the level
was different than the pool ids of the things which the level consisted of.
The schedule info can be wrong in such cases because the direction is stored in the level
and the columns from the level pool may be different than the thing pool.